third_party/WebKit/LayoutTests/http/tests/inspector/cache-storage/cache-live-update-cache-content.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="cache-storage-test.js"></script>
<script>

function test() {
    var cacheStorageModel = InspectorTest.mainTarget.model(SDK.ServiceWorkerCacheModel);
    cacheStorageModel.enable();

    function errorAndExit(error) {
        if (error)
            InspectorTest.addResult(error);
        InspectorTest.completeTest();
    }

    function isMarkedDirty() {
        var view = UI.panels.resources._sidebar.cacheStorageListTreeElement.childAt(0)._view;
        InspectorTest.addResult("Cache marked dirty = " + view._needsRefresh.visible());
    }

    function main() {
        InspectorTest.clearAllCaches()
            .then(InspectorTest.dumpCacheTree)
            .then(InspectorTest.createCache.bind(this, "testCache1"))
            .then(InspectorTest.dumpCacheTree)
            .then(isMarkedDirty)
            .then(InspectorTest.addCacheEntry.bind(this, "testCache1", "http://fake.request.com/1", "OK"))
            .then(InspectorTest.addResult.bind(this, "Added entry"))
            .then(isMarkedDirty)
            .then(InspectorTest.dumpCacheTree)
            .then(isMarkedDirty)
            .then(InspectorTest.deleteCacheEntry.bind(this, "testCache1", "http://fake.request.com/1"))
            .then(InspectorTest.addResult.bind(this, "Deleted entry"))
            .then(isMarkedDirty)
            .then(InspectorTest.dumpCacheTree)
            .then(isMarkedDirty)
            .then(InspectorTest.clearAllCaches)
            .then(InspectorTest.completeTest)
            .catch(errorAndExit);
    }

    InspectorTest.waitForCacheRefresh(main);
}
</script>
</head>
<body onload="runTest()">
<p>Tests that cache content is marked dirty if it is modified.</p>
</body>
</html>
